+++ src/main/java/shopping/domain/user/Password.java
@@ -1,7 +1,5 @@
 package shopping.domain.user;
 
-import shopping.infrastructure.PasswordEncoder;
-
 import javax.persistence.Embeddable;
 
 @Embeddable
