+++ src/main/java/shopping/domain/user/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package shopping.infrastructure;
+package shopping.domain.user;
 
 public interface PasswordEncoder {
     String encode(final String password);
