+++ src/main/java/shopping/infrastructure/SHA256PasswordEncoder.java
@@ -1,6 +1,7 @@
 package shopping.infrastructure;
 
 import org.springframework.stereotype.Component;
+import shopping.domain.user.PasswordEncoder;
 import shopping.exception.ErrorType;
 import shopping.exception.ShoppingException;
